+++ CHANGELOG.md
@@ -11,6 +11,10 @@ details.
 
 ---
 
+## [Unreleased]
+### Added
+- Support to process rules from multiple locations in`checkrules.js`.
+
 ## [1.2.0] - 2017-01-07
 ### Added
 - All published rules as of ESLint 3.13.0